Adds multitexturing support, FreeImage library for image loading, two more test objects and some textures

+++ AVTProject1/AVTProject1/src/Shader.cpp
@@ -136,7 +136,7 @@ namespace AVTEngine
 
 	void Shader::setFloat(const GLchar* field, GLfloat value)
 	{
-		GLfloat loc = glGetUniformLocation(shaderID, field);
+		GLint loc = glGetUniformLocation(shaderID, field);
 		glUniform1f(loc, value);
 	}
